cuda, fit: fix two -Werror build failures in the new cache code
The CUDA CI job builds with fatal warnings and both of these are in code
added by this PR, so the job has been failing at 'Build with CMake'.

1. ggml-cuda.cu, VMM pool OOM retry:

     error: comparison between 'enum cudaError_enum' and 'enum cudaError'
       [-Werror=enum-compare]

   cuMemCreate is a driver-API call returning CUresult, while
   cudaErrorMemoryAllocation is a runtime-API cudaError_t. The comment said
   the per-vendor alias kept this portable, but that only holds on HIP and
   MUSA, whose vendor headers map cuMemCreate onto the runtime error type
   (hip.h:142 and hip.h:85). On CUDA the two are distinct enums that happen
   to share the value 2, so the logic worked wherever it compiled at all.
   Uses CUDA_ERROR_OUT_OF_MEMORY on CUDA and keeps the runtime constant on
   HIP/MUSA.

2. common/fit.cpp, layer-footprint overflow guard:

     error: comparison of integer expressions of different signedness:
       'uint64_t' and 'long int' [-Werror=sign-compare]

   Both operands are already guarded positive on the preceding lines, so the
   right-hand side is cast to uint64_t, preserving the guard's meaning.

Verified on aarch64 Linux + CUDA 13.0 with -DLLAMA_FATAL_WARNINGS=ON: a
keep-going build reports 0 errors and produces all binaries; test-moe-cache,
test-moe-cache-fit and test-turbo-quant all exit 0.

+++ common/fit.cpp
@@ -1239,7 +1239,7 @@ static void common_params_fit_impl(
                 for (const llama_moe_tensor_info & tensor : moe_tensors) {
                     if (tensor.layer >= 0 && tensor.layer < total_layers &&
                         tensor.n_expert > 0 && tensor.expert_size > 0 &&
-                        (uint64_t)tensor.n_expert <= INT64_MAX / (int64_t)tensor.expert_size) {
+                        (uint64_t)tensor.n_expert <= (uint64_t)(INT64_MAX / (int64_t)tensor.expert_size)) {
                         layer_bytes[tensor.layer] +=
                             (int64_t)tensor.n_expert * (int64_t)tensor.expert_size;
                     }
